Класс Graph.java: * Написан метод для вычисления зависимостей граней и узлов. * Дописан конструктор. * Переписаны метод: List<Edge> getEdgesByNode(Node node) -> getNodeById(long nodeId), containsNode(Node node) -> containsNode(long nodeId)
Класс Node.java:
* Созданы поля lat и lon, для хранения географического расположения.

Класс OsmParser.java:
* Возымел метод convertToGraph()

+++ src/main/java/mapObjects/Edge.java
@@ -9,7 +9,7 @@ public class Edge {
     private final double speed;
     private double length;
 
-    public Edge(long id, List<Long> idNodes, int speed, double length) {
+    public Edge(long id, List<Long> idNodes, double speed, double length) {
         this.id = id;
         this.idNodes = idNodes;
         this.speed = speed;
